+++ examples/01-blog/Blog/Type/Scalar/EmailType.php
@@ -18,7 +18,8 @@ class EmailType extends ScalarType
     public function serialize($value): string
     {
         if (! $this->isEmail($value)) {
-            throw new SerializationError('Cannot represent value as email: ' . Utils::printSafe($value));
+            $notEmail = Utils::printSafe($value);
+            throw new SerializationError("Cannot represent value as email: {$notEmail}");
         }
 
         return $value;
@@ -27,7 +28,8 @@ class EmailType extends ScalarType
     public function parseValue($value): string
     {
         if (! $this->isEmail($value)) {
-            throw new Error('Cannot represent value as email: ' . Utils::printSafeJson($value));
+            $notEmail = Utils::printSafeJson($value);
+            throw new Error("Cannot represent value as email: {$notEmail}");
         }
 
         return $value;
@@ -38,7 +40,7 @@ class EmailType extends ScalarType
         // Note: throwing GraphQL\Error\Error vs \UnexpectedValueException to benefit from GraphQL
         // error location in query:
         if (! $valueNode instanceof StringValueNode) {
-            throw new Error('Query error: Can only parse strings got: ' . $valueNode->kind, [$valueNode]);
+            throw new Error("Query error: Can only parse strings got: {$valueNode->kind}", [$valueNode]);
         }
 
         $value = $valueNode->value;