+++ src/Utils/Value.php
@@ -35,7 +35,7 @@ class Value
      *
      * @phpstan-return CoercedValue|CoercedErrors
      */
-    public static function coerceValue($value, InputType $type, ?array $path = null): array
+    public static function coerceInputValue($value, InputType $type, ?array $path = null): array
     {
         if ($type instanceof NonNull) {
             if ($value === null) {
@@ -45,7 +45,7 @@ class Value
             }
 
             // @phpstan-ignore-next-line wrapped type is known to be input type after schema validation
-            return self::coerceValue($value, $type->getWrappedType(), $path);
+            return self::coerceInputValue($value, $type->getWrappedType(), $path);
         }
 
         if ($value === null) {
@@ -80,7 +80,7 @@ class Value
                 $errors = [];
                 $coercedValue = [];
                 foreach ($value as $index => $itemValue) {
-                    $coercedItem = self::coerceValue(
+                    $coercedItem = self::coerceInputValue(
                         $itemValue,
                         $itemType,
                         [...$path ?? [], $index]
@@ -99,7 +99,7 @@ class Value
             }
 
             // Lists accept a non-list value as a list of one.
-            $coercedItem = self::coerceValue($value, $itemType);
+            $coercedItem = self::coerceInputValue($value, $itemType);
 
             return isset($coercedItem['errors'])
                 ? $coercedItem
@@ -124,7 +124,7 @@ class Value
         foreach ($fields as $fieldName => $field) {
             if (array_key_exists($fieldName, $value)) {
                 $fieldValue = $value[$fieldName];
-                $coercedField = self::coerceValue(
+                $coercedField = self::coerceInputValue(
                     $fieldValue,
                     $field->getType(),
                     [...$path ?? [], $fieldName],