+++ src/Validator/Rules/PossibleTypeExtensions.php
@@ -112,7 +112,7 @@ class PossibleTypeExtensions extends ValidationRule
             case NodeKind::INPUT_OBJECT_TYPE_DEFINITION:
                 return NodeKind::INPUT_OBJECT_TYPE_EXTENSION;
             default:
-                throw new InvariantViolation("Unexpected definition kind: {$kind}");
+                throw new InvariantViolation("Unexpected definition kind: {$kind}.");
         }
     }
 
@@ -133,7 +133,7 @@ class PossibleTypeExtensions extends ValidationRule
                 return NodeKind::INPUT_OBJECT_TYPE_EXTENSION;
             default:
                 $unexpectedType = Utils::printSafe($type);
-                throw new InvariantViolation("Unexpected type: {$unexpectedType}");
+                throw new InvariantViolation("Unexpected type: {$unexpectedType}.");
         }
     }
 
@@ -153,7 +153,7 @@ class PossibleTypeExtensions extends ValidationRule
             case NodeKind::INPUT_OBJECT_TYPE_EXTENSION:
                 return 'input object';
             default:
-                throw new InvariantViolation("Unexpected extension kind: {$kind}");
+                throw new InvariantViolation("Unexpected extension kind: {$kind}.");
         }
     }
 }